General Changes
concert.html: started the changeover to the -QTY+ method
ticketing.py: Now checks for quantities and linked products
timedFunctions.py: moved directories around a bit
payment.html: added a stripe integrated payment page
ticketRouting.py: moved all ticket related routing to this file
urls.py: Added references to ticketRouting.py
view.py: Added references to payment page
