+++ kelvin/kelvin/settings.py
@@ -18,7 +18,7 @@ CONCERT_LIST = [
     {
          'soldOut': False,
          'date': '18th November 2023',
-         'tickets': [{'ticketLabel': "Standard Ticket", 'ticketID': 'price_1OFVVXDysBLU7VPvWnC7jH1Z'}, {'ticketLabel': "Concession Ticket", 'ticketID': 'price_1OFVVXDysBLU7VPvJhWYmQ1D'}, {'ticketLabel': "Standing Ticket", 'ticketID': 'price_1OFhKCDysBLU7VPvXpb95h3Q'}]
+         'tickets': [{'ticketLabel': "Standard Ticket", 'ticketID': 'price_1OFVVXDysBLU7VPvWnC7jH1Z'}, {'ticketLabel': "Concession Ticket", 'ticketID': 'price_1OFVVXDysBLU7VPvJhWYmQ1D'}, {'ticketLabel': "Restricted View Ticket", 'ticketID': 'price_1OFhKCDysBLU7VPvXpb95h3Q'}]
     }
 ]
 
@@ -42,6 +42,8 @@ STRIPE_PUBLIC_SECRET = os.environ['STRIPE_PUBLIC_SECRET']
 STRIPE_ENDPOINT_SECRET = os.environ['STRIPE_ENDPOINT_SECRET']
 
 TICKETS_PASSWORD = os.environ['tickets_password']
+
+COOKIE_DATA = {}
 
 # STRIPE SETTINGS ######
 REDIRECT_DOMAIN = 'http://yukisuter.pythonanywhere.com'
