+++ kelvin/website/templates/website/concerts_old.html
@@ -316,7 +316,7 @@ Thank you to everyone who came to our <b>Spring 2023: Romance</b> concert.  <br>
             height="450" frameborder="0" allowtransparency="true" allow="encrypted-media"></iframe>
         <iframe src="https://embed.music.apple.com/gb/playlist/kelvin-ensemble-spring-2023/pl.u-J9jJSDRzpXXv"
             height="450px" frameborder="0"
-            sandbox="allow-forms allow-popups allow-same-origin allow-scripts allow-top-navigation-by-user-activation"
+            sandbox="allow-forms allow-popups allow-same-origin allow-ticketing allow-top-navigation-by-user-activation"
             allow="autoplay *; encrypted-media *;"
             style="width: 100%; max-width: 900px; overflow: hidden; border-radius: 10px; background: transparent;"></iframe>
     </div>
