+++ kelvin/website/templates/website/includes/scripts.html
@@ -10,7 +10,7 @@
 <script src="{% static 'website/vendor/jquery-easing/jquery.easing.min.js' %}"></script>
 <script src="{% static 'website/vendor/scrollreveal/scrollreveal.min.js' %}"></script>
 
-<!-- Custom scripts for this template -->
+<!-- Custom ticketing for this template -->
 <script src="{% static 'website/vendor/creative/creative.js' %}"></script>
 
 <script>
